+++ python/arkimet/scan/__init__.py
@@ -1,6 +1,9 @@
+import _arkimet
 from . import grib
 from . import bufr
 from . import vm2
 from . import odimh5
 
-__all__ = ("grib", "bufr", "vm2", "odimh5")
+Scanner = _arkimet.scan.Scanner
+
+__all__ = ("Scanner", "grib", "bufr", "vm2", "odimh5")